# backend/openapi.json
{
  "components": {
    "schemas": {
      "AgentName": {
        "enum": [
          "genesis",
          "cipher",
          "chronos",
          "forensiq",
          "spectra",
          "atlas"
        ],
        "title": "AgentName",
        "type": "string"
      },
      "AgentResult": {
        "additionalProperties": false,
        "description": "The only result format permitted from a forensic agent.",
        "properties": {
          "agent": {
            "$ref": "#/components/schemas/AgentName"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "title": "Evidence",
            "type": "array"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "title": "Findings",
            "type": "array"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "processing_time_ms": {
            "minimum": 0.0,
            "title": "Processing Time Ms",
            "type": "integer"
          },
          "version": {
            "pattern": "^v\\d+(\\.\\d+){0,2}$",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "agent",
          "confidence",
          "processing_time_ms",
          "version"
        ],
        "title": "AgentResult",
        "type": "object"
      },
      "ArtifactType": {
        "enum": [
          "plain_text",
          "email",
          "document",
          "resume",
          "voice",
          "image",
          "certificate"
        ],
        "title": "ArtifactType",
        "type": "string"
      },
...
      "CreateIdentityGenomeRequest": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "owner_id": {
            "format": "uuid",
            "title": "Owner Id",
            "type": "string"
          }
        },
        "required": [
          "owner_id",
          "display_name"
        ],
        "title": "CreateIdentityGenomeRequest",
        "type": "object"
      },
      "CreateInvestigationRequest": {
        "additionalProperties": false,
        "properties": {
          "artifact_reference": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Artifact Reference",
            "type": "string"
          },
          "artifact_type": {
            "$ref": "#/components/schemas/ArtifactType"
          },
          "identity_genome_id": {
            "format": "uuid",
            "title": "Identity Genome Id",
            "type": "string"
          }
        },
        "required": [
          "identity_genome_id",
          "artifact_type",
          "artifact_reference"
        ],
        "title": "CreateInvestigationRequest",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "pattern": "^[A-Z0-9_]+$",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id",
          "timestamp"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "EvidenceItem": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Category",
            "type": "string"
          },
          "code": {
            "pattern": "^[a-z0-9_]+$",
            "title": "Code",
            "type": "string"
          },
          "source_reference": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Source Reference",
            "type": "string"
          },
          "weight": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "code",
          "category",
          "weight",
          "source_reference"
        ],
        "title": "EvidenceItem",
        "type": "object"
      },
      "Finding": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "pattern": "^[a-z0-9_]+$",
            "title": "Code",
            "type": "string"
          },
          "evidence_codes": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Codes",
            "type": "array"
          },
          "severity": {
            "pattern": "^(info|low|medium|high|critical)$",
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "code",
          "severity"
        ],
        "title": "Finding",
        "type": "object"
      },
      "IdentityGenomeResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "owner_id": {
            "format": "uuid",
            "title": "Owner Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "owner_id",
          "display_name",
          "created_at",
          "updated_at"
        ],
        "title": "IdentityGenomeResponse",
        "type": "object"
      },
      "IdentityProfileResponse": {
        "properties": {
          "average_word_count": {
            "title": "Average Word Count",
            "type": "number"
          },
          "embedding_count": {
            "title": "Embedding Count",
            "type": "integer"
          },
          "identity_genome_id": {
            "format": "uuid",
            "title": "Identity Genome Id",
            "type": "string"
          },
          "sample_count": {
            "title": "Sample Count",
            "type": "integer"
          },
          "unique_token_count": {
            "title": "Unique Token Count",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "identity_genome_id",
          "sample_count",
          "unique_token_count",
          "average_word_count",
          "embedding_count",
          "updated_at"
        ],
        "title": "IdentityProfileResponse",
        "type": "object"
      },
      "IngestTextRequest": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 500000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "source_label": {
            "default": "text-sample",
            "maxLength": 120,
            "minLength": 1,
            "title": "Source Label",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "IngestTextRequest",
        "type": "object"
      },
      "InvestigationResponse": {
        "properties": {
          "artifact_reference": {
            "title": "Artifact Reference",
            "type": "string"
          },
          "artifact_type": {
            "$ref": "#/components/schemas/ArtifactType"
          },
          "case_number": {
            "title": "Case Number",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "identity_genome_id": {
            "format": "uuid",
            "title": "Identity Genome Id",
            "type": "string"
          },
          "opened_at": {
            "format": "date-time",
            "title": "Opened At",
            "type": "string"
          },
          "risk_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/InvestigationStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "verdict": {
            "$ref": "#/components/schemas/Verdict"
          }
        },
        "required": [
          "id",
          "case_number",
          "identity_genome_id",
          "artifact_type",
          "artifact_reference",
          "status",
          "verdict",
          "risk_level",
          "opened_at",
          "updated_at"
        ],
        "title": "InvestigationResponse",
        "type": "object"
      },
      "InvestigationRunResponse": {
        "properties": {
          "agents": {
            "items": {
              "$ref": "#/components/schemas/AgentResult"
            },
            "title": "Agents",
            "type": "array"
          },
          "investigation": {
            "$ref": "#/components/schemas/InvestigationResponse"
          },
          "risk": {
            "$ref": "#/components/schemas/RiskAssessment"
          }
        },
        "required": [
          "investigation",
          "agents",
          "risk"
        ],
        "title": "InvestigationRunResponse",
        "type": "object"
      },
      "InvestigationStatus": {
        "enum": [
          "queued",
          "investigating",
          "completed",
          "failed"
        ],
        "title": "InvestigationStatus",
        "type": "string"
      },
      "RiskAssessment": {
        "properties": {
          "rationale_codes": {
            "items": {
              "type": "string"
            },
            "title": "Rationale Codes",
            "type": "array"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "verdict": {
            "$ref": "#/components/schemas/Verdict"
          }
        },
        "required": [
          "risk_level",
          "verdict",
          "rationale_codes"
        ],
        "title": "RiskAssessment",
        "type": "object"
      },
      "RiskLevel": {
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "RiskLevel",
        "type": "string"
      },
      "RunTextInvestigationRequest": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 500000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "RunTextInvestigationRequest",
        "type": "object"
      },
      "Verdict": {
        "enum": [
          "pending",
          "authentic",
          "inconclusive",
          "impersonation_confirmed"
        ],
        "title": "Verdict",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "TrustDNA API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/identity-genomes": {
      "post": {
        "operationId": "create_identity_genome_api_v1_identity_genomes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIdentityGenomeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityGenomeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Create Identity Genome",
        "tags": [
          "Identity Genome"
        ]
      }
    },
    "/api/v1/identity-genomes/{genome_id}": {
      "get": {
        "operationId": "get_identity_genome_api_v1_identity_genomes__genome_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "genome_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Genome Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityGenomeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Identity Genome",
        "tags": [
          "Identity Genome"
        ]
      }
    },
    "/api/v1/identity-genomes/{genome_id}/samples/text": {
      "post": {
        "operationId": "ingest_text_sample_api_v1_identity_genomes__genome_id__samples_text_post",
        "parameters": [
          {
            "in": "path",
            "name": "genome_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Genome Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestTextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Ingest Text Sample",
        "tags": [
          "Identity Genome"
        ]
      }
    },
    "/api/v1/investigations": {
      "post": {
        "operationId": "open_investigation_api_v1_investigations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvestigationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Open Investigation",
        "tags": [
          "Investigations"
        ]
      }
    },
    "/api/v1/investigations/{investigation_id}": {
      "get": {
        "operationId": "get_investigation_api_v1_investigations__investigation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "investigation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Investigation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Investigation",
        "tags": [
          "Investigations"
        ]
      }
    },
    "/api/v1/investigations/{investigation_id}/run-text": {
      "post": {
        "operationId": "run_text_investigation_api_v1_investigations__investigation_id__run_text_post",
        "parameters": [
          {
            "in": "path",
            "name": "investigation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Investigation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunTextInvestigationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Run Text Investigation",
        "tags": [
          "Investigations"
        ]
      }
    },
...
  }
}